WeekTest_Insert
I changed the type in week id to be a double so the id of a week is
<number>.<month> so the test has 2.3 as the id to mean the second week
in the third month.

+++ src/ModelLayer/Week.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class Week {
 	
-	private int id;
+	private double id;
 	private int number; 
 	private int month; 
 	private int year; 
@@ -24,11 +24,11 @@ public class Week {
 		this.stateList = stateList;
 	}
 	
-	public int getId() {
+	public double getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(double id) {
 		this.id = id;
 	}
